requirements.txt
Flask
requests
beautifulsoup4
python-dotenv # For loading .env file
openai # For OpenAI API access